# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/employees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of employee profiles with optional pagination. Admin role required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List employee profiles (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.EmployeeProfileResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/employees/{userID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the details of a specific employee profile. Admin role required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get employee profile (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID of the employee",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.EmployeeProfileResponse"
                        }
                    },
                    "400": {
                        "description": "invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "profile not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates fields in an employee's profile. Admin role required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update employee profile (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID of the employee",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.PatchEmployeeProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.EmployeeProfileResponse"
                        }
                    },
                    "400": {
                        "description": "invalid user ID, invalid JSON or invalid fields",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "profile not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ima_diplom-backend_internal_handler_dto.EmployeeProfileResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "corporate_email": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "dismissal_date": {
                    "type": "string"
                },
                "employee_code": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gdp_training_history": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "medical_book_scan_url": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "special_zone_access": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ima_diplom-backend_internal_handler_dto.PatchEmployeeProfileRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "corporate_email": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "dismissal_date": {
                    "type": "string"
                },
                "employee_code": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gdp_training_history": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hire_date": {
                    "type": "string"
                },
                "medical_book_scan_url": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "special_zone_access": {
                    "type": "boolean"
                }
            }
        },
...
    }
}